[Sim] Build target boot image for simulator
The simulator requires a target profile and boot image to run and
correctly simulate target tests so add a build step to generate a
target profile and boot image on the host if the simulator is being
built.

Also allow the bootclasspath-location to take a prefix so that
non-device paths (e.g.: the target or host out directory) to the
bootclasspath files can be used.

Author: Chris Jones <christopher.jones@arm.com>

Test: export ART_USE_SIMULATOR=true
      test.py --host --target
Change-Id: I8c9f2a1c2e3f14bbc48bc3f2eef18a682ce5557b