Implement AVX2 instructions in x64 assembler
This patch implements the AVX2 instructions required for supporting all existing HVEC instructions. This is a building block to enable AVX2-based 256-bit vectorization to improve codec performance.

x86 has overlapping FP and vector registers, so xmm/ymm is the same register (xmm is the lower half of the larger ymm register). In order to determine if a register is xmm or ymm, vector registers were introduced in a separate patch that included the vector length.

Test: art/test.py -b --host
      Run the above test with and without "avx,avx2" features
Bug: 202101983
Change-Id: Ibaaed0b57e5ed3d99eb963ab46f66332c9b3c0c6
Signed-off-by: Vinay Kompella <vinay.kompella@intel.com>