Introduce the notion of vector registers
When handling parallel moves, it is important to use the correct
register width, xmm or ymm. Without this patch, it is difficult to
determine whether to use xmm or ymm. The assumption to use a ymm when a graph has SIMD will technically succeed for correctness, but can cause severe performance regressions (~50%) as switching between AVX and SSE is expensive without using a vzeroupper. This patch addresses these issues by storing the vector length in the registers.

On ARM, x86 vector registers are extensions of FP registers x86 and, as such, are one and the same. They cannot be used simultaneously without clobbering the other. This patch focuses mainly these on overlapping FP/vector registers. Vector registers are modeled here as FP registers with a vector length > 0.

The use of vector registers will be driven by vector length from loop vectorization, therefore vector codegen need not request for vector registers via LocationsBuilder. The vector length is saved in the locations during register allocation. Implicitly the allocated FP registers contain this info, making them vector registers.

Note: Risc-v has exclusive set of vector registers.

Test: art/test.py -b --host
      Run the above test with and without "avx,avx2" features
Bug: 202101983
Change-Id: I9bea8fb13ffb3cb77a13753aae2356c49fba5837
Signed-off-by: Vinay Kompella <vinay.kompella@intel.com>