Add PaletteGetPriorityMapping
ART sometimes needs access to the mapping itself, so that it can,
for example, use an interpolated priority for system daemons.

aosp/3287177 contains the current version of the proposed client code.

Add PaletteClientTest for GetPriorityMapping.

This is in preparation for allowing ART to further manipulate thread
priorities. See go/thread-priorities-in-android and b/330444460 and
its many friends.

Driveby fix for PaletteDebugStoreGetString() in palette_fake.cc.
It was fixed in the near-copy under system, but not here.

Test: Treehugger
Test: Treehugger in conection with aosp/3287177 and aosp/3286067
Change-Id: Ic7ba204b2d25cb82c9db01b21389b958ae527793