Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/99c3d2ce-036d-4231-9995-65ec96dc02b0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3a25b7f6b218f9935679f1d8c34656dd717d9d48
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com