Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/24b305dd-46fa-4815-b670-3378e1984c2f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I21c02dc2943c04dbb0e9c50ad277ecf610c3a565
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com