Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/707e6f88-9271-49a3-81d9-a83d7dc964b7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iafd571195fb8e6a5f99193fc898434a5def6d305
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com