Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a2861336-f0b2-4ac5-b71f-8207c899cdd8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I391075d716fe1a76fd8d4f9fd7c8b74c4024afa5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com