Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a6659d62-292f-48c0-965e-a853e9065a7e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6df48097ca80f773a46414d127cc72a9a61a8724
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com