Revert "Clean up host APEX support."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1238ac2d-d804-43de-a051-70850b63467a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I920b6c0836b825f76c031ff3c4bf7fb09ad8e79e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com