Revert "Use uint64_t instead of size_t for black-dense calculations"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bf9a10d6-7caf-43f7-aad4-14ad34b0d4e9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0e340961990b5fbeb397945332cef1e5734497d9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com