Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/474918f2-faa3-4661-ae5a-559ab61dc0a4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic18cb69299a17362199c410d6949349bc9756543
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com