Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c3b5214c-614e-4201-b715-69e3c7008185). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If452b613136dae6ae845ed1ba89814c938a6599d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com