Revert "Use uint64_t instead of size_t for black-dense calculations"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/353d4df0-03be-4e58-8558-5556ec8e3f24). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ida57e1035e650e85e05699b226846ca990a2c143
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com