Revert "Clean up host APEX support."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9441e6a7-9488-48f4-b517-79f0ea8426dc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I21cffcf71e1c1fccb66393aac2e894c356732102
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com