Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8fd3aab2-591f-481c-a7c4-0d7ca3578ef1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I28e4a18a6508decdd41f14b0a63d69ff3f8254eb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com