Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3dc7eb7f-2364-4541-afbf-865ecf13e81c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I95fdd89cc2ef50d5cfe0dfc62745473d98420d23
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com