Revert "cleanup: Remove extra SetGraph calls"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/220f06f4-dbc8-4fc9-b078-8aaaa0fed057). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If19d046be832022eaf491a0aa7e6ca1060ef2df8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com