Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/220f06f4-dbc8-4fc9-b078-8aaaa0fed057). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifea2d3798389c8a6c5692bb2ef19d4ef35aea7dd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com