Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3a3139fe-fdc3-4a28-9353-63b8139d6f52). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ied35a0256e262e0df6b27c1b222923878ecc0557
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com