Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c36edef3-74b8-49cb-832c-533ae7713b93). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I83fe5f7693ed93125e84fa429bda92b78ce2c814
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com