Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ef22a4d3-5a90-4c64-af0e-4e58e30fe82f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I646eff6dabf48e70d74ea4016568d91b9cf4c7ff
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com