Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d2949026-fa1e-4146-997c-f9209f8674cd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6a6e7886053e6f179742d521fc0137c17db22caf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com