Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/75b8ffcb-280f-40ad-b9a9-b0fc8e81fbaf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2874466b74286c192fe4769dd8682b81c12a4e01
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com