Revert "verifier: Re-introdce `VERIFY_ERROR_NO_FIELD`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cfad6fdd-ac6d-4928-8142-151b17a1f688). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idfe35ebeec5a417ca5178e7ebf6dab69e1ad0fe2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com