Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eef4e793-d46c-4e79-a889-d13295e96f8b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibe7f3967bf088c1a5e4e463919dcd091bf80ee7a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com