Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/972d888c-c9e5-402e-a5eb-89a4707b13ef). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I49c25f7f05f36429a02d002a15cb5259978cf0db
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com