Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d00214f2-2c0f-4074-b5a6-988fd28d4a99). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0c71fed57c9fef0406d8dacbcb0463788343adc6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com