Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ee1ef2ee-3fd8-493d-95d8-4cb7b003cb62). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I88851e977034b79f52eacf3ed01b61b862a980b5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com