Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a1508b9-01b7-4668-be02-40acd94916dc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7460bea4fce2c71fc42b6874446d7f606ee5d7f8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com