Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cc85295e-4df5-4d8f-87b9-8996e461c558). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7482959b131cb74859e03231c123a4ea30bc2761
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com