Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/697af2a4-7964-48b4-bbf0-66e98fcd9dca). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I20a98125eafd8241b4d37e978a73104481dc060e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com