Revert "Add more WellKnownClasses fields for boxing/unboxing."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2d8fc052-ead2-4606-b043-b0adcfaa4c66). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I45abd27f8cb0f08395351430a0bf404ba9d61dde
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com