Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cf97b715-c3a4-4bc2-b7b0-6612a1a565f1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2cbfc47145a2962393c957e611271747ef180d16
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com