Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ef03b3ea-b044-4834-b4ee-5be412b1caec). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8e22456dfe79cae542b4ac748f675450f82dcffd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com