Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/67caa60d-f4c7-44c5-a8fb-a2e811cf0337). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iffbfc98c1a371443fb0bb627e9a2c623ba022a2f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com