art: x86_64: Disable uffd gc if emulating large page size
When emulating 16KB page size on x86_64, the kernel presents a 16kB page
size to userspace. However the kernel and hardware still operates at a
4KB page size granularity.

This mean that even though the mimumun size of memory that userspace can
request is 16384 (from mmap); faults still happen at at 4KB granularity
in the kernel.

This is inherently incompatible with UFFD GC, which was made the default
for V. Since UFFD necessarily needs to operate on PTEs (4096) but from
the client's perspective in userspace the page size is 16384.

Disable the UFFD GC if emulating 16KB (larger than 4KB) page sizes on
x86_64.

This change only affects cuttlefish and android studio x86_64 16KB
targets. On all other targets this is effectively a no-op.

Bug: 379001861
Bug: 377361489
Test: Manually tested Android studio emulator
Change-Id: Ieeceb05879f8864595a7166d796aee6037df4ba5
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>