art: Use kernel page size for pagemap index calculation
Android emulates 16KB page size on x86_64 to faciliate development
and testing of 16KB apps.

x86_64 emulates a 16kB page size to userspace but under the hood the
kernel and hardware still operates on 4KB page sizes.

This discrepancy between the kernel and user space page size usually
is transparent to userspace; but for operations like reading
/proc/pid/pagemap and mincore() which looks at PTEs we need to
handle the conversion from userspace page size to kernel page size.

This is effectively a no-op on all other targets.

Bug: 384919556
Flag: EXEMPT bugfix
Test: Presubmit
Change-Id: I2f5fa59fb79a5813aa975b6bae1bb6464a4e9da0