Revert "Add test.java.lang.runtime.SwitchBootstrapsTest to run-l..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f562001e-19a4-460e-86e2-faba8b2d9bd1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I638965ddf2dac29fea3fd4b81a65b2286db591a6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com