Revert "Add test.java.lang.runtime.SwitchBootstrapsTest to run-l..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/99db7dc0-d5b8-4a4c-865f-00a6f0c6b9a1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8c56cdb31e861abe80b9a5de3447040764d51ed2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com