Revert "Add back a test that was missed from a previous CL."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a85c2e4f-75ba-4fa4-b991-087bb8a7361a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7eaa1ccaa0ad061b2a5411eef8cf9de823111091
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com