Revert "qemu: Update to Ubuntu 24.04 LTS (Noble Numbat)"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/378034b5-ba2a-4603-8437-74f7d1d27055). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I80ad5dc0d5a748b406a3da0c61ed2fef7e5d8469
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com