Revert "Add some folks to the tools/ahat/OWNERS file."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b7e03cc0-c95c-4c35-bce5-f62a8e6cf1b5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I102a46606beab4e5b0f5a6fcf6d4efaaa249cfae
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com