Revert "verifier: Inline `Check.*Op*()`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6ad40da2-aa4d-4b66-887e-3d990bd065f6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I28e4d41c8d0d9aa1602628e5fa75af5e55c5ea54
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com