Revert "verifier: Inline `Check.*Op*()`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/913f004e-1cac-4d42-bd68-00a5a797df13). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8d611b1467785aa15c8f5fcdf6b3a9fea357f0d0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com