Revert "verifier: Inline `Check.*Op*()`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/180a9f15-49a1-4e3e-b73a-91b8bc8929cb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibbf27266203ad903f9032c2b5a07beac6cec3169
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com