Revert "verifier: Inline `Check.*Op*()`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d7d68d6b-c71f-4868-9f20-7aab755f413f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1e30fc7abe45288b8f55e6037b42edf37dcfcadf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com