Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/06bc17ef-5497-4b70-b4a4-44373572bc56). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1b52fbc247dd6506bec903a2190af2001e86d707
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com