Revert "cleanup: Remove extra SetGraph calls"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d2603259-12ba-4f7d-b7ab-631c9564addb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibbcf08edf10dc290c62461d117db7cb194bb0925
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com