Revert "Reduce hotness on interpreter lookups."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d2603259-12ba-4f7d-b7ab-631c9564addb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iad16be6739ab1456a726ccee8b2af0c4aff00a40
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com