Revert "verifier: Inline `Check.*Op*()`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a0d4d6e-2cac-439b-8523-35c03a40762f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I845c8ff5ff54421372c4949d5bee96359381f254
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com