Revert "verifier: Inline `Check.*Op*()`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/061b64ab-228a-4a7b-8563-0bed058a586c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3770f4598055d958f5dcf1b7c5d34c2a1bcc2fdf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com