Revert "verifier: Inline `Check.*Op*()`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1318111f-de3c-4761-a82f-2624fe491478). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I79e3429aad7a5abb0e67901ab5e46db90c8aea03
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com