Revert "Rename `GetBootImageVarHandleField()`, drop "Boot"."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/20f92d14-7913-458b-ac04-9a8244060c10). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia86536fc09c9c89455975dcaf5417d10e457ea26
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com