Revert "verifier: Inline `Check.*Op*()`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8df3c1f6-70e0-4ea5-a001-7db4e7f82535). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic9aba5a3da1be316dcba5d5d433b7b272002b18d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com