Revert "Rename `GetBootImageVarHandleField()`, drop "Boot"."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b68bd4ca-8880-4279-8629-c7b166cb82ce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I39962df4f8d9265e014ddd340aed61ff03453ef5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com