Revert "verifier: Inline `Check.*Op*()`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0e59374c-2ddd-470b-abe9-872fc5a7747b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ief9eef13b4675de714b9eb2eeeb864196197e9e6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com