Don't allow duplicate methods or fields
To mark methods as duplicate they need to have the
same name and signature. For fields, same name and
type. This is different than the Java requirement
as DEX files allow e.g. to have an "int foo" and
"long foo" fields.

The uploaded DEX file has a duplicate method, and
the duplicate fields check is done as a drive-by.

Bug: 391852979
Test: m test-art-host-gtest-art_runtime_tests64
Change-Id: I4b8c20d62d79b363bf1b017f2f46efbef01e9432