Update AssignVTableIndexes regarding duplicate methods
We allow for duplicate methods within the same class
in the direct methods or the virtual methods, as we
will register the first one and ignore the second one.
Caveat: we do not allow duplicates across those lists,
so this means that class A can have two `foo` direct
OR virtual methods, but it cannot have one `foo` direct
and one `foo` virtual.

Bug: 391852979
Test: m test-art-host-gtest-art_runtime_tests64
Change-Id: I4b8c20d62d79b363bf1b017f2f46efbef01e9432