Revert "verifier: Reduce duplication for `kVerifierDebug`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/414efdd2-2365-4c7b-b345-07d1720e2701). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id76248711109334252566108be9d729fdd851224
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com