Revert "Handle invoke-virtual targeting invokeExact gracefully."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/954eae9e-2106-4774-8c24-36349231c569). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic4f0a1373dbaf0b0fc1f668e3b4233b25f4ea495
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com