Revert "verifier: Speed up failure recording."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5bc7b385-fc53-47cd-a7c2-dc8fa4aae6fe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib77b5509373f3ece12b860fd5d532e156b423fd7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com