Improve class fuzzer initialization
Add libsigchain as a library, which lets us remove the no-sig-chain
argument and the sigaltstack workaround. To do this, we import
libart-runtime-for-test_static_defaults instead of the regular runtime
one to avoid having libsigchain_fake as a libary.

Remove explicit WellKnownClasses::Init since RunEarlyRootClinits already
does it.

Move TransitionFromRunnableToSuspended to the end of the initialization
which is where we would do Runtime::Start(). This also lets us remove
the ScopedObjectAccess.

Bug: 352721437
Test: m test-art-host-gtest-art_runtime_tests64
Test: Follow art/tools/fuzzer/README.md for libart_verify_classes_fuzzer
Test: Follow art/tools/fuzzer/README.md for libart_verify_classes_fuzzerd
Change-Id: I7d7b8bb78e13cd0c49172f92ca8293fbb806f01e