Revert "Adjust art tests to java.lang changes."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a16a5d52-c550-4c36-aab7-820af4b6801d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3bab2fbd1302ecc18fc3d00b8d5c6ac41d8553e7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com