Revert "Enable lazy release on CC for some targets."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/20ebff61-c587-439a-b66a-d0e2dd6a0f60). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 390354527
Change-Id: I7736722fac0c271513e3da5a6614b50877c54d9f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com