Revert "verifier: Do not mark registers as conflict for return."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a237150-6007-40e7-99f0-e8abafed89c3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I212c87af2548dd52e959587d26f4fe8bdaf75a70
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com