Revert "Minimal hiddenapibypass breakage demo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d299ef58-02b6-4686-8b06-b57b40495f2a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 390288768
Change-Id: I99a7ccd4b4850ddfa37dba722fb63bc6b2e53b76
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com