Revert "Minimal hiddenapibypass breakage demo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/911f19b0-e3f2-4962-b2a9-b37f730b7094). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 390292480
Change-Id: Ic5ca729ec3a4df434f55526972770df52195b9e0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com