Revert "Adjust art tests to java.lang changes."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bf5e4bc7-5129-493d-8073-96041c2ff856). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I36c65c2b6db152c74aa2f277d9dc35b1c1391bfa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com