Revert "Revert "Always use an array in the DexCache for ArtField..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/00685330-ea90-4225-9b63-3d7545aef75e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If2df5fa2786b681e42b3869b13b4a467a68b8a1d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com