Stop using MAP_32BIT for mapping in <4GB on x86_64
MAP_32BIT flag limits to first 2GB address space as per manpage. However,
as per the kernel source code it seems to be only 1GB, thereby limiting
available address space for mapping in <4GB on x86_64.

Also, fixed an issue with CMC GC and free-list large-object space
(somehow got revealed only with this CL) wherein a nullptr gets added
to its live-bitmap and then a subsequent Sweep() on the space fails an
assertion.

Bug: 384045833
Test: art/test/testrunner/testrunner.py --host --64
Change-Id: If5f696648a9d55bd08d4f39ee7a25dea285b89dc