Revert "Merge sFields and iFields."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2e2d02e7-4f4a-481c-be4a-b23e37cd486e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7a4eba4adc92fbc26f8ba5dc32219e646a9293f4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com