Revert "Make CMC GC generational"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/25911b94-3e6b-4035-ab6d-555c8308ea17). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391531096
Change-Id: I5b6c5d07e2b8cee27c1edcf0cb47feb6dfe4cea5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com