Revert "Enable Generational CMC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/25911b94-3e6b-4035-ab6d-555c8308ea17). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391531096
Change-Id: I167579c90c504b78587bac3d177d170d68bb0ffa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com