Revert "Optimizing: Allow moving `HCondition` to use site."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/87005a73-a980-47d3-8dac-15067e917030). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 0
Bug: 0
Change-Id: Ia31fef49dcb063a9011348a886ddbba9fb96c4f9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com