Revert "Optimizing: Generate `HSelect` if there are more phis..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6f4b84e7-b3bf-4de1-b9a2-7c8d8f04094e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I54dbac4964f414575e5aa4fceeef3b72cf168dff
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com