Revert "Add a one-pass baseline compiler for arm64."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8bfb97b0-271b-4159-9877-9eebec803c69). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib0300c0474e067f9cdb49160e21ea4fa7c88cf27
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com