Revert "Simplify PrimitiveArray<T>::Memcpy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c1f08dbb-479e-4033-877b-0f6349793814). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8aacd8233e121555691c2fa8fb79bab143a8e259
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com