Revert "Avoid deflating monitors in the zygote space."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/83e9941d-84ec-4669-be37-472810442da1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4930c5107404622845a55ab69afc9cf23ba5d97b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com