Revert "Avoid deflating monitors in the zygote space."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/08b0c181-67b1-4cd8-a891-9243c01b96dc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ife496578402029fb4adb2dfe14fc771c6cfd2334
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com