Revert "Avoid deflating monitors in the zygote space."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5141a3b6-5ccd-4fe2-9378-5f54243e9595). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I526823682462ca0b605a27bc1331306105d7f1ed
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com