Revert "Avoid deflating monitors in the zygote space."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a38cd44d-7753-44fe-8d97-cc1fd583ace8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie33b45681c701d729eb921f3c7e21358be1ceb5e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com