Revert "Avoid deflating monitors in the zygote space."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c371e4bc-3d36-4ba8-af0b-a0a8ac4adaa8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6f079e4cbf2f9d909adc762fa9b04bd5da717d2d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com