Revert "Hide most symbols in `ClassLinker`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5e2a9ea7-6950-4ba6-bb52-c8a9e7a38340). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 393113705
Change-Id: I254cc67d25e9198198e18a684c5c531014298842
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com