Revert "Add option to write zeros instead of sparse."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8081f96d-99e9-4c57-99dd-c705137d3e54). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8eeaaec4b30983f273ac881909107d85576fa635
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com