Revert "Introduce an "external" variant of libarttest(d).so and ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/efdfe5fb-4013-4bf0-81ea-36b44d8dc89b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1ec99aa0bdce2e785947894ab9878b32d7f51732
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com