Revert "Enable Generational CMC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7e4cbfa7-e3ca-43d1-873f-eb450b418043). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8a05b6b8c2cdc3a158dce30cf3de899d52da6493
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com