Revert "Avoid deflating monitors in the zygote space."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c99936b0-5303-4043-91d3-2f385f86549d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib6780b553e6c7db46dc963a3c02168759f884e2f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com