Revert "Use libarttest(d)_external to fix 674-hiddenapi, and ena..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a3e37408-f07c-4ea8-9058-4f7048f9a4d3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 392821354
Change-Id: I0b268d1898e177f19de65f2de2a2c9ec5ac09668
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com