Revert "Allow ART internal libs to load libs in NATIVELOADER_DEF..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a3e37408-f07c-4ea8-9058-4f7048f9a4d3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 392821354
Change-Id: I226d5d2dd245ba5a14d2744b74d00ec13492b01a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com