Revert "Allow ART internal libs to load libs in NATIVELOADER_DEF..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5c2e2a62-2134-4d9a-87ba-ddba883c0315). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 392821354
Change-Id: I42cdefac684493e5e235c6843c0b2e837f94941f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com