Revert "Use libarttest(d)_external to fix 674-hiddenapi, and ena..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5c2e2a62-2134-4d9a-87ba-ddba883c0315). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 392821354
Change-Id: Ice3fb90ba01eee669dd84081c192eb15ebd65a3e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com