Revert "Fix the loading of libarttest(d).so in 150-loadlibrary a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5c2e2a62-2134-4d9a-87ba-ddba883c0315). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 392821354
Change-Id: I027512702025577e759682150df71fda37fe312e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com