Revert "Add option to write zeros instead of sparse."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d5ee4f66-4c8a-40d7-8e13-d0023221625b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1a09516a39496d410bd2a8492b95cbf9276bfc88
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com