Revert "Add option to write zeros instead of sparse."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e0d24143-af58-4671-b147-9c54e0ffe235). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie12ca7b1081a389a6ebc546bd0e5fd95b460e814
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com