Revert "Add option to write zeros instead of sparse."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a24f8e98-6c26-4dce-8176-9c00d5b9c89f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I67235a8c9b2fc9dbf7e591f46f435dd4f2b17549
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com